Seperate dom from scripts
