+++ test/Customer-test.js
@@ -1,12 +1,6 @@
-// import chai from 'chai';
-// const expect = chai.expect;
 import { assert } from 'chai';
 
 import Customer from '../src/classes/Customer';
-// import Hotel from '../src/Hotel';
-// import sampleCustomerData from '../src/sample-data/sample-customer-data';
-// import sampleRoomData from '../src/sample-data/sample-room-data';
-// import sampleBookingData from '../src/sample-data/sample-booking-data';
 
 describe('Customer', () => {
   let customer;
